generate names scaffold
name is a one to one relationship with pictures. Names belongs to pics
but only a name id belongs to a picture. the names table is only to be
obtained by a user not manipulated

migrate: name is string and has an id(int), and index method
serilaized: id and name can be seen
model: