batch import/seed data
wrote a script for importing names and pics with
correct cooresponding ids to avoid mix ups.
ORDER MATTERS when importing the rows in the script.

Couldnt find a way to import the columns simulateously so the solution
that worked was splitting them into 2 seperate csv files. The data is
seeded correctly in the corresponing tables